[capture] switch to mostly 16bits interleaved capture

+++ testFunc/test_adc.cpp
@@ -29,7 +29,7 @@ void testOne(adc_smp_rate one, adc_prescaler two,int sc)
    adc->prepareDMASampling (one,two);            
    adc->startDMASampling (1024);
    SampleSet    set;
-   uint32_t data[256];
+   uint16_t data[256];
    set.data=data;
    //bool r=adc->getSamples(set);
